requirements.txt
anyio==4.2.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
aws-cdk-asset-awscli-v1==2.2.202 ; python_version >= "3.11" and python_version < "4.0"
aws-cdk-asset-kubectl-v20==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
aws-cdk-asset-node-proxy-agent-v6==2.0.1 ; python_version >= "3.11" and python_version < "4.0"
aws-cdk-lib==2.122.0 ; python_version >= "3.11" and python_version < "4.0"
boto3==1.34.26 ; python_version >= "3.11" and python_version < "4.0"
botocore==1.34.26 ; python_version >= "3.11" and python_version < "4.0"
cattrs==23.2.3 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
cloudpickle==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
constructs==10.3.0 ; python_version >= "3.11" and python_version < "4.0"
contextlib2==21.6.0 ; python_version >= "3.11" and python_version < "4.0"
dill==0.3.7 ; python_version >= "3.11" and python_version < "4.0"
docker==7.0.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.95.2 ; python_version >= "3.11" and python_version < "4.0"
google-pasta==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.6 ; python_version >= "3.11" and python_version < "4.0"
importlib-metadata==6.11.0 ; python_version >= "3.11" and python_version < "4.0"
importlib-resources==6.1.1 ; python_version >= "3.11" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
jsii==1.94.0 ; python_version >= "3.11" and python_version < "4.0"
jsonschema==4.21.1 ; python_version >= "3.11" and python_version < "4.0"
multiprocess==0.70.15 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.26.3 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
pathos==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==4.25.2 ; python_version >= "3.11" and python_version < "4.0"
psutil==5.9.8 ; python_version >= "3.11" and python_version < "4.0"
pydantic==1.10.14 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
sagemaker==2.204.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.66.1 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.22.0 ; python_version >= "3.11" and python_version < "4.0"
zipp==3.17.0 ; python_version >= "3.11" and python_version < "4.0"
